requirements.txt
# Brand Campaign Generator — Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# AI
anthropic==0.40.0

# PDF Parsing
PyMuPDF==1.25.1

# Utilities
python-multipart==0.0.12    # FastAPI file uploads
pydantic==2.10.3
httpx==0.28.1               # Async HTTP (used by anthropic)
requests==2.32.3            # OpenRouter/OpenAI image and text calls
Pillow==11.0.0              # Image composition and logo/text overlays
python-dotenv==1.0.1        # Loads API keys from .env
